requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[argon2]
python-jose
email-validator
pydantic